requirements.txt
# Pygame 2.x 以上 (仮想キー描画やスケーリング機能のため)
pygame-ce>=2.3.0

# ブラウザ実行用のランタイム（開発環境・ビルド用）
pygbag>=0.8.0

# 非同期処理用（標準ライブラリですが、環境明示のため）
asyncio